Fix pink terrain artifacts and optimize viewport rendering
- Fixed an integer division bug in `SingleMeshTerrainRenderer` where `UVTileSize` was 0, causing the shader to fail texture sampling.
- Updated `Terrain.shader` to use vertex colors (placeholders) instead of debug pink when textures are loading or dimensions are invalid.
- Implemented a multi-layer caching system (`_cellCache`, `_metadataCache`) and eliminated all per-frame allocations during camera movement.
- Optimized mesh updates by using a viewport-sized sliding window and manual bounds calculation.
- Ensured Z-buffer safety during progressive loading by using `discard` for transparent pixels.

+++ Assets/Scripts/World/SingleMeshTerrainRenderer.cs
@@ -221,7 +221,7 @@ namespace Fodinae.Assets.Scripts.World
                 HasTileGroup = MapManager.Instance.TryGetTileGroup(type, out int gid), TileGroupId = gid,
                 MinimapColor = MapManager.Instance.GetCellMinimapColor(type), Animation = config.Animation,
                 AnimationSpeed = config.AnimationSpeed, AtlasRect = atlasRect, AtlasIndex = atlasIndex,
-                UVTileSize = atlases.Count > atlasIndex ? RenderingConstants.CELL_SIZE / atlases[atlasIndex].Size : 0,
+                UVTileSize = atlases.Count > atlasIndex ? (float)RenderingConstants.CELL_SIZE / atlases[atlasIndex].Size : 0,
                 IsTextureReady = atlasRect.z > 0.0001f
             };
             _metadataCache[type] = meta;
